refactor: extract types and mapping
