+++ web-components/src/components/inputs/new/ConfirmationCode/ConfirmationCode.mapping.ts
@@ -1,0 +1,23 @@
+import { InputProps } from './ConfirmationCode.types';
+
+export const propsMap: { [key: string]: InputProps } = {
+  alpha: {
+    type: 'text',
+    inputMode: 'text',
+    pattern: '[a-zA-Z]{1}',
+  },
+
+  alphanumeric: {
+    type: 'text',
+    inputMode: 'text',
+    pattern: '[a-zA-Z0-9]{1}',
+  },
+
+  numeric: {
+    type: 'tel',
+    inputMode: 'numeric',
+    pattern: '[0-9]{1}',
+    min: '0',
+    max: '9',
+  },
+};
