+++ web-components/src/components/modal/EmailConfirmationModal/EmailConfirmationModal.stories.tsx
@@ -1,6 +1,5 @@
 /* eslint-disable react-hooks/rules-of-hooks */
 import { useState } from 'react';
-import { Link } from '@mui/material';
 import { action } from '@storybook/addon-actions';
 import { Meta, StoryObj } from '@storybook/react';
